+++ eva/configuration/constants.py
@@ -33,3 +33,4 @@ INDEX_DIR = "index"
 CACHE_DIR = "cache"
 DATASET_DATAFRAME_NAME = "dataset"
 DB_DEFAULT_URI = "sqlite:///{}/eva_catalog.db".format(EVA_DEFAULT_DIR.resolve())
+S3_DOWNLOAD_DIR = "s3_downloads"