+++ eva/optimizer/rules/rules.py
@@ -43,7 +43,6 @@ from eva.plan_nodes.show_info_plan import ShowInfoPlan
 if TYPE_CHECKING:
     from eva.optimizer.optimizer_context import OptimizerContext
 
-from eva.configuration.configuration_manager import ConfigurationManager
 from eva.optimizer.operators import (
     Dummy,
     LogicalApplyAndMerge,
@@ -51,6 +50,7 @@ from eva.optimizer.operators import (
     LogicalCreateIndex,
     LogicalCreateMaterializedView,
     LogicalCreateUDF,
+    LogicalDelete,
     LogicalDrop,
     LogicalDropUDF,
     LogicalExplain,
@@ -77,6 +77,7 @@ from eva.optimizer.operators import (
 from eva.plan_nodes.create_index_plan import CreateIndexPlan
 from eva.plan_nodes.create_plan import CreatePlan
 from eva.plan_nodes.create_udf_plan import CreateUDFPlan
+from eva.plan_nodes.delete_plan import DeletePlan
 from eva.plan_nodes.drop_plan import DropPlan
 from eva.plan_nodes.drop_udf_plan import DropUDFPlan
 from eva.plan_nodes.faiss_index_scan_plan import FaissIndexScanPlan
@@ -89,7 +90,6 @@ from eva.plan_nodes.limit_plan import LimitPlan
 from eva.plan_nodes.load_data_plan import LoadDataPlan
 from eva.plan_nodes.orderby_plan import OrderByPlan
 from eva.plan_nodes.rename_plan import RenamePlan
-from eva.plan_nodes.sample_plan import SamplePlan
 from eva.plan_nodes.seq_scan_plan import SeqScanPlan
 from eva.plan_nodes.storage_plan import StoragePlan
 from eva.plan_nodes.union_plan import UnionPlan
@@ -207,63 +207,6 @@ class EmbedProjectIntoGet(Rule):
         )
 
         yield new_get_opr
-
-
-# For nested queries
-
-
-class EmbedFilterIntoDerivedGet(Rule):
-    def __init__(self):
-        pattern = Pattern(OperatorType.LOGICALFILTER)
-        pattern_get = Pattern(OperatorType.LOGICALQUERYDERIVEDGET)
-        pattern_get.append_child(Pattern(OperatorType.DUMMY))
-        pattern.append_child(pattern_get)
-        super().__init__(RuleType.EMBED_FILTER_INTO_DERIVED_GET, pattern)
-
-    def promise(self):
-        return Promise.EMBED_FILTER_INTO_DERIVED_GET
-
-    def check(self, before: Operator, context: OptimizerContext):
-        # nothing else to check if logical match found return true
-        return True
-
-    def apply(self, before: LogicalFilter, context: OptimizerContext):
-        predicate = before.predicate
-        ld_get = before.children[0]
-        new_opr = LogicalQueryDerivedGet(
-            alias=ld_get.alias,
-            predicate=predicate,
-            target_list=ld_get.target_list,
-            children=ld_get.children,
-        )
-        yield new_opr
-
-
-class EmbedProjectIntoDerivedGet(Rule):
-    def __init__(self):
-        pattern = Pattern(OperatorType.LOGICALPROJECT)
-        pattern_get = Pattern(OperatorType.LOGICALQUERYDERIVEDGET)
-        pattern_get.append_child(Pattern(OperatorType.DUMMY))
-        pattern.append_child(pattern_get)
-        super().__init__(RuleType.EMBED_PROJECT_INTO_DERIVED_GET, pattern)
-
-    def promise(self):
-        return Promise.EMBED_PROJECT_INTO_DERIVED_GET
-
-    def check(self, before: Operator, context: OptimizerContext):
-        # nothing else to check if logical match found return true
-        return True
-
-    def apply(self, before: LogicalProject, context: OptimizerContext):
-        target_list = before.target_list
-        ld_get = before.children[0]
-        new_opr = LogicalQueryDerivedGet(
-            alias=ld_get.alias,
-            predicate=ld_get.predicate,
-            target_list=target_list,
-            children=ld_get.children,
-        )
-        yield new_opr
 
 
 class CacheFunctionExpressionInFilter(Rule):
@@ -394,7 +337,7 @@ class PushDownFilterThroughJoin(Rule):
             new_join_node.append_child(right)
 
         if rem_pred:
-            new_join_node.join_predicate = conjuction_list_to_expression_tree(
+            new_join_node._join_predicate = conjuction_list_to_expression_tree(
                 [rem_pred, new_join_node.join_predicate]
             )
 
@@ -488,9 +431,10 @@ class PushDownFilterThroughApplyAndMerge(Rule):
 
         # If we have partial predicate make it the root
         root_node = apply_and_merge
-        if rem_pred:
-            root_node = LogicalFilter(predicate=rem_pred)
-            root_node.append_child(apply_and_merge)
+        assert rem_pred is None
+        # if rem_pred:
+        #    root_node = LogicalFilter(predicate=rem_pred)
+        #    root_node.append_child(apply_and_merge)
 
         yield root_node
 
@@ -540,16 +484,10 @@ class CombineSimilarityOrderByAndLimitToFaissIndexScan(Rule):
 
         # Check if predicate exists on table.
         def _exists_predicate(opr):
-            if isinstance(opr, LogicalFilter):
-                return True
-            elif isinstance(opr, LogicalGet):
+            if isinstance(opr, LogicalGet):
                 return opr.predicate is not None
-            elif isinstance(opr, LogicalQueryDerivedGet):
-                return opr.predicate is not None
-            exists_predicate = False
-            for child in opr.children:
-                exists_predicate |= _exists_predicate(child)
-            return exists_predicate
+            # LogicalFilter
+            return True
 
         if _exists_predicate(sub_tree_root.opr):
             return
@@ -768,6 +706,22 @@ class LogicalInsertToPhysical(Rule):
         yield after
 
 
+class LogicalDeleteToPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALDELETE)
+        super().__init__(RuleType.LOGICAL_DELETE_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_DELETE_TO_PHYSICAL
+
+    def check(self, before: Operator, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalDelete, context: OptimizerContext):
+        after = DeletePlan(before.table_ref, before.where_clause)
+        yield after
+
+
 class LogicalLoadToPhysical(Rule):
     def __init__(self):
         pattern = Pattern(OperatorType.LOGICALLOADDATA)
@@ -780,20 +734,9 @@ class LogicalLoadToPhysical(Rule):
         return True
 
     def apply(self, before: LogicalLoadData, context: OptimizerContext):
-        # Configure the batch_mem_size.
-        # We assume the optimizer decides the batch_mem_size.
-        # ToDO: Experiment heuristics.
-
-        batch_mem_size = 30000000  # 30mb
-        config_batch_mem_size = ConfigurationManager().get_value(
-            "executor", "batch_mem_size"
-        )
-        if config_batch_mem_size:
-            batch_mem_size = config_batch_mem_size
         after = LoadDataPlan(
             before.table_info,
             before.path,
-            batch_mem_size,
             before.column_list,
             before.file_options,
         )
@@ -812,21 +755,10 @@ class LogicalUploadToPhysical(Rule):
         return True
 
     def apply(self, before: LogicalUpload, context: OptimizerContext):
-        # Configure the batch_mem_size.
-        # We assume the optimizer decides the batch_mem_size.
-        # ToDO: Experiment heuristics.
-
-        batch_mem_size = 30000000  # 30mb
-        config_batch_mem_size = ConfigurationManager().get_value(
-            "executor", "batch_mem_size"
-        )
-        if config_batch_mem_size:
-            batch_mem_size = config_batch_mem_size
         after = UploadPlan(
             before.path,
             before.video_blob,
             before.table_info,
-            batch_mem_size,
             before.column_list,
             before.file_options,
         )
@@ -849,41 +781,14 @@ class LogicalGetToSeqScan(Rule):
         # Configure the batch_mem_size. It decides the number of rows
         # read in a batch from storage engine.
         # ToDO: Experiment heuristics.
-
-        batch_mem_size = 30000000  # 30mb
-        config_batch_mem_size = ConfigurationManager().get_value(
-            "executor", "batch_mem_size"
-        )
-        if config_batch_mem_size:
-            batch_mem_size = config_batch_mem_size
         after = SeqScanPlan(None, before.target_list, before.alias)
         after.append_child(
             StoragePlan(
                 before.table_obj,
-                batch_mem_size=batch_mem_size,
                 predicate=before.predicate,
                 sampling_rate=before.sampling_rate,
             )
         )
-        yield after
-
-
-class LogicalSampleToUniformSample(Rule):
-    def __init__(self):
-        pattern = Pattern(OperatorType.LOGICALSAMPLE)
-        pattern.append_child(Pattern(OperatorType.DUMMY))
-        super().__init__(RuleType.LOGICAL_SAMPLE_TO_UNIFORMSAMPLE, pattern)
-
-    def promise(self):
-        return Promise.LOGICAL_SAMPLE_TO_UNIFORMSAMPLE
-
-    def check(self, before: Operator, context: OptimizerContext):
-        return True
-
-    def apply(self, before: LogicalSample, context: OptimizerContext):
-        after = SamplePlan(before.sample_freq)
-        for child in before.children:
-            after.append_child(child)
         yield after