+++ test/integration_tests/test_pytorch.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-import sys
 import unittest
+from test.markers import windows_skip_marker
 from test.util import (
     copy_sample_images_to_upload_dir,
     copy_sample_videos_to_upload_dir,
@@ -46,7 +46,7 @@ class PytorchTest(unittest.TestCase):
         ua_detrac = f"{EVA_ROOT_DIR}/data/ua_detrac/ua_detrac.mp4"
         mnist = f"{EVA_ROOT_DIR}/data/mnist/mnist.mp4"
         actions = f"{EVA_ROOT_DIR}/data/actions/actions.mp4"
-        asl_actions = f"{EVA_ROOT_DIR}/data/actions/computer_asl.avi"
+        asl_actions = f"{EVA_ROOT_DIR}/data/actions/computer_asl.mp4"
         meme1 = f"{EVA_ROOT_DIR}/data/detoxify/meme1.jpg"
         meme2 = f"{EVA_ROOT_DIR}/data/detoxify/meme2.jpg"
 
@@ -63,7 +63,7 @@ class PytorchTest(unittest.TestCase):
         file_remove("ua_detrac.mp4")
         file_remove("mnist.mp4")
         file_remove("actions.mp4")
-        file_remove("computer_asl.avi")
+        file_remove("computer_asl.mp4")
 
         execute_query_fetch_all("DROP TABLE IF EXISTS Actions;")
         execute_query_fetch_all("DROP TABLE IF EXISTS Mnist;")
@@ -105,11 +105,14 @@ class PytorchTest(unittest.TestCase):
     def test_should_run_pytorch_and_asl(self):
         execute_query_fetch_all(Asl_udf_query)
         select_query = """SELECT FIRST(id), ASLActionRecognition(SEGMENT(data))
-                        FROM Asl_actions SAMPLE 5 GROUP BY '16f';"""
+                        FROM Asl_actions
+                        SAMPLE 5
+                        GROUP BY '16f';"""
         actual_batch = execute_query_fetch_all(select_query)
 
         res = actual_batch.frames
 
+        self.assertEqual(len(res), 1)
         for idx in res.index:
             self.assertTrue("computer" in res["aslactionrecognition.labels"][idx])
 
@@ -130,7 +133,7 @@ class PytorchTest(unittest.TestCase):
         self.assertEqual(len(actual_batch), 5)
 
     @pytest.mark.torchtest
-    @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
+    @windows_skip_marker
     def test_should_run_pytorch_and_ocr(self):
         create_udf_query = """CREATE UDF IF NOT EXISTS OCRExtractor
                   INPUT  (frame NDARRAY UINT8(3, ANYDIM, ANYDIM))
@@ -226,7 +229,7 @@ class PytorchTest(unittest.TestCase):
         self.assertTrue(np.array_equal(img, similar_data))
 
     @pytest.mark.torchtest
-    @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
+    @windows_skip_marker
     def test_should_run_ocr_on_cropped_data(self):
         create_udf_query = """CREATE UDF IF NOT EXISTS OCRExtractor
                   INPUT  (text NDARRAY STR(100))
@@ -249,7 +252,7 @@ class PytorchTest(unittest.TestCase):
         self.assertTrue(res["ocrextractor.scores"][2][0] > 0.9)
 
     @pytest.mark.torchtest
-    @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
+    @windows_skip_marker
     def test_should_run_detoxify_on_text(self):
         create_udf_query = """CREATE UDF IF NOT EXISTS OCRExtractor
                   INPUT  (text NDARRAY STR(100))